Split big index.js file into multiple smaller files
- Follows the plan laid out in
https://github.com/jupyterhub/binderhub/issues/777
- Only non-trivial move is to cleanup the `_config` call to
  use the more modern 'fetch' API rather than the older
  XMLHttpRequest API

Fixes https://github.com/jupyterhub/binderhub/issues/777
