+++ binderhub/static/js/src/constants.js
@@ -1,0 +1,3 @@
+
+export const BASE_URL = $('#base-url').data().url;
+export const BADGE_BASE_URL = $('#badge-base-url').data().url;
